requirements.txt
python-telegram-bot
apscheduler
pandas
openpyxl
python-dotenv
pytz
ultralytics
opencv-python-headless

